% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AD_pheno}
\alias{AD_pheno}
\title{AD_pheno}
\format{
An object of class \code{data.frame} with 47 rows and 5 columns.
}
\source{
\code{
#### How you made AD_pheno ####
usethis::use_data(AD_pheno, overwrite = TRUE)
}
}
\usage{
data("AD_pheno")
}
\description{
Description of AD_pheno.
}
\keyword{datasets}
