% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{AD_unionCounts}
\alias{AD_unionCounts}
\title{AD_unionCounts}
\format{
An object of class \code{list} of length 2.
}
\source{
\code{
#### How you made AD_unionCounts ####
usethis::use_data(AD_unionCounts, overwrite = TRUE)
}
}
\usage{
data("AD_unionCounts")
}
\description{
Description of AD_unionCounts
}
\keyword{datasets}
