% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AD_unionCounts_bulk.R
\docType{data}
\name{AD_unionCounts_bulk}
\alias{AD_unionCounts_bulk}
\title{AD entorhinal cortex H3K27ac read counts for consensus peaks.}
\format{
A counts per million matrix with 251345 rows and 47 variables, with
rows labelling consensus peaks and columns labelling samples.
}
\source{
Generated using data from \url{https://www.nature.com/articles/s41593-018-0253-7}
and \url{https://github.com/nottalexi/brain-cell-type-peak-files/tree/master/H3K27ac}
}
\usage{
AD_unionCounts_bulk
}
\description{
A dataset containing read counts generated using H3K27ac data from entorhinal cortex
for consensus peaks generated by the CHAS-MF function UnionPeaks(). This dataset is
used as an example for the CHAS-MF tutorial
}
\keyword{datasets}
