% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHAS.R
\name{AnnotBulkPeaks}
\alias{AnnotBulkPeaks}
\title{Annotate bulk H3K27ac peaks with celltypes}
\usage{
AnnotBulkPeaks(bulkPeaks, celltype, celltypeOverlap)
}
\arguments{
\item{bulkPeaks}{A data frame of bulk H3K27ac peaks in BED format, with the first
column containing the chromosome number, the second column containing the start
position of the peak, the third column containing the end position of the peak, and
the fourth column containing the peak identifier.}

\item{celltype}{A character specifying the celltype to be annotated to the bulk peaks.}

\item{celltypeOverlap}{The data frame containing the overlapping bulk/celltype H3K27ac
peaks from the function BulkCelltypeOverlap().}
}
\value{
The bulk H3K27ac peaks data frame with the celltype annotation.
}
\description{
This function annotates each bulk tissue H3K27ac peak with the celltype(s) it overlaps.
It takes as input bulk H3K27ac peaks in bed format, the name of the celltype to be
used in the annotation, and the output of the function BulkCelltypeOverlap().
}
