% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHAS.R
\name{BulkCelltypeOverlap}
\alias{BulkCelltypeOverlap}
\title{Overlap bulk H3K272ac profiles with cell sorted H3K27ac data}
\usage{
BulkCelltypeOverlap(bulkPeaks, celltypePeaks)
}
\arguments{
\item{bulkPeaks}{A data frame of bulk H3K27ac peaks in BED format, with the first
column containing the chromosome number, the second column containing the start
position of the peak, the third column containing the end position of the peak, and
the fourth column containing the peak identifier.}

\item{celltypePeaks}{A data frame of celltype H3K27ac peaks in BED format, with the
first column containing the chromosome number, the second column containing the start
position of the peak, the third column containing the end position of the peak, and
the fourth column containing the peak identifier.}
}
\value{
A data frame of overlapping bulk/celltype H3K27ac peaks.
}
\description{
This function takes as input bulk H3K27ac peaks and celltype sorted H3K27ac peaks,
and will generate a data frame of overlapping bulk/celltype peaks.
}
