% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_median_and_variability.R
\name{calculate_median_and_variability}
\alias{calculate_median_and_variability}
\title{Calculate Median and Variability}
\usage{
calculate_median_and_variability(tpm, samples)
}
\arguments{
\item{tpm}{The TPM data frame.}

\item{samples}{The samples data frame.}
}
\value{
A list containing the median and variability data frames.
}
\description{
This function calculates the median and variability for reference counts.
}
