% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_tpm.R
\name{calculate_tpm}
\alias{calculate_tpm}
\title{Calculate TPM}
\usage{
calculate_tpm(counts, peaks, signature)
}
\arguments{
\item{counts}{The counts data frame.}

\item{peaks}{The peaks data frame.}

\item{signature}{The signature vector.}
}
\value{
A data frame of TPM values.
}
\description{
This function calculates the Transcripts Per Million (TPM) for the given counts, peaks, and signature.
}
