% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CelltypeProportion.R
\name{CelltypeProportion}
\alias{CelltypeProportion}
\title{MF add-on Function 2: Calculate cell type proportions}
\usage{
CelltypeProportion(consensus, refSamples)
}
\arguments{
\item{consensus}{The output list from Function 1}

\item{refSamples}{A data frame containing cell type annotations for reference samples
the first column contains the sample ID used in refCounts,
they should be in the same order as the columns in reference counts
the second column contains the corresponding cell type for that sample}
}
\value{
A list containing the following:
\link{1} a list: numbers of signature peaks for each cell type
\link{2} data frame: EPIC-predicted cell-type proportions
}
\description{
This function calculates cell type proportions using matrix factorisation
}
\details{
This function takes two inputs:
(1) the output list from Funtoin 1
(2) cell type annotations for reference samples

Prerequisite:
install the EPIC R package from https://github.com/GfellerLab/EPIC

This function has three steps:
(1) calculate the median and variability for reference counts
(2) select cell-type-specific signature peaks
for a give peak, compare the median counts among cell types
if the maximum counts is at least 5 times larger than the second largest
then this peak is added to the signature peak set
(3) run matrix factorisation with EPIC
}
