% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHAS.R
\name{CelltypeScore}
\alias{CelltypeScore}
\title{Calculate a celltype-specific histone acetylation score}
\usage{
CelltypeScore(cpm, MinOverlapPeaks)
}
\arguments{
\item{cpm}{A counts per million matrix with rows labeling peaks and columns labeling
samples.}

\item{MinOverlapPeaks}{The output data frame from the function MinOverlap() containing
bulk H3K27ac peaks which are celltype-specific and overlap > x\% of a celltype h3K27ac peak.}
}
\value{
Celltype-specific histone acetylation scores for each sample.
}
\description{
This function takes as input a counts per million (cpm) matrix and the output of function
MinOverlap(), and will calculate a celltype-specific histone acetylation score for
each sample in the cpm matrix.
}
