% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHAS.R
\name{CelltypeSpecificPeaks}
\alias{CelltypeSpecificPeaks}
\title{Identify celltype-specific H3K27ac peaks}
\usage{
CelltypeSpecificPeaks(bulkPeaks, celltype)
}
\arguments{
\item{bulkPeaks}{A data frame of bulk H3K27ac peaks in BED format, with the first
column containing the chromosome number, the second column containing the start
position of the peak, the third column containing the end position of the peak, and
the fourth column containing the peak identifier.}

\item{celltype}{A character specifying the celltype of interest.}
}
\value{
A data frame containing bulk H3K27ac peaks which are celltype-specific.
}
\description{
This function identifies which bulk H3K27ac peaks are specific to a celltype.
Bulk H3K27ac peaks which overlap more than one cell type are annotated to "Multiple",
and those that do not overlap any of the celltypes investigated are annotated to
"Other".
}
