% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConsensusPeaks.R
\name{ConsensusPeaks}
\alias{ConsensusPeaks}
\title{MF add-on Function 1: Find Consensus Peaks}
\usage{
ConsensusPeaks(bulkPeaks, bulkCounts, refPeaks, refCounts, bedtools_path)
}
\arguments{
\item{bulkPeaks}{A data frame of bulk H3K27ac peaks in BED format
the first column contains the chromosome number, such as 'chr1'
the second column contains the start position of the peak
the third column contains the end position of the peak
the fourth column contains the peak identifier}

\item{bulkCounts}{A counts matrix for bulk data
rows represent peaks, using the same identifier as bulkPeaks
columns represent bulk samples}

\item{refPeaks}{A data frame of reference H3K27ac peaks in BED format
the first column contains the chromosome number, such as 'chr1'
the second column contains the start position of the peak
the third column contains the end position of the peak
the fourth column contains the peak identifier}

\item{refCounts}{A counts matrix for reference data
rows represent peaks, using the same identifier as refPeaks
columns represent cell-type-specific samples, can can contain
either: 1 sample for each cell type
or: 3 samples for each cell type}

\item{bedtools_path}{The path to where bedtools is installed
on Mac, this can be checked by runing "\% which bedtools" in Bash}
}
\value{
A list containing the following:
\link{1} data frame: bulk normalised counts
\link{2} data frame: reference normalised counts
}
\description{
This function identifies the consensus peaks and their counts in each bulk and reference
samples by switching IDs
}
\details{
This function takes four inputs:
bulk: peak location, raw counts
reference: peak location, raw counts

This function has four steps:
(1) calculate length-normalised cpm from raw counts
(2) find consensus peaks in bulk and reference
(3) assign new identifiers to consensus peaks
(4) filter the length-normalised cpm matrices to contain only the consensus peaks
}
