% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EntorhinalCortex_AD_H3K27ac_counts.R
\docType{data}
\name{EntorhinalCortex_AD_H3K27ac_counts}
\alias{EntorhinalCortex_AD_H3K27ac_counts}
\title{Entorhinal cortex H3K27ac read counts.}
\format{
A counts per million matrix with 183353 rows and 47 variables, with
rows labelling peaks and columns labelling samples.
}
\source{
Generated using data from \url{https://www.nature.com/articles/s41593-018-0253-7}
}
\usage{
EntorhinalCortex_AD_H3K27ac_counts
}
\description{
A dataset containing read counts generated using H3K27ac data from entorhinal cortex
from AD cases and controls.
}
\keyword{datasets}
