% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EntorhinalCortex_AD_H3K27ac_peaks.R
\docType{data}
\name{EntorhinalCortex_AD_H3K27ac_peaks}
\alias{EntorhinalCortex_AD_H3K27ac_peaks}
\title{Entorhinal cortex H3K27ac peaks with hg38 coordinates.}
\format{
A data frame with 70923 rows and 4 variables:
\describe{
\item{CHR}{chromosome number}
\item{Start}{Start location of peak in BP}
\item{End}{End location of peak in BP}
\item{Peak}{Peak name}
...
}
}
\source{
Generated using data from \url{https://www.nature.com/articles/s41593-018-0253-7}
}
\usage{
EntorhinalCortex_AD_H3K27ac_peaks
}
\description{
A dataset containing entorhinal cortex H3K27ac peaks from AD and control brains
with hg38 coordinates in bed file format.
}
\keyword{datasets}
