% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgl_H3K27ac_hg19.R
\docType{data}
\name{mgl_H3K27ac_hg19}
\alias{mgl_H3K27ac_hg19}
\title{Microglia H3K27ac peaks with hg19 coordinates.}
\format{
A data frame with 48836 rows and 4 variables:
\describe{
\item{CHR}{chromosome number}
\item{Start}{Start location of peak in BP}
\item{End}{End location of peak in BP}
\item{Peak}{Peak name}
...
}
}
\source{
\url{https://github.com/nottalexi/brain-cell-type-peak-files/tree/master/H3K27ac}
}
\usage{
mgl_H3K27ac_hg19
}
\description{
A dataset containing microglia H3K27ac peaks with hg19 coordinates in bed
file format.
}
\keyword{datasets}
