% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHAS.R
\name{MinOverlapOfPeaks}
\alias{MinOverlapOfPeaks}
\title{Identify celltype-specific H3K27ac peaks which overlap > x\% of the celltype}
\usage{
MinOverlapOfPeaks(CelltypeOverlap, CelltypeSpecificPeaks, x)
}
\arguments{
\item{CelltypeOverlap}{The output data frame from the function BulkCelltypeOverlap()
containing the overlapping bulk/celltype H3K27ac peaks.}

\item{CelltypeSpecificPeaks}{The output data frame from the function CelltypeSpecificPeaks()
containing bulk H3K27ac peaks which are specific to a celltype.}

\item{x}{A character specifying the minimum percentage overlap required between the
bulk H3K27ac peak and the celltype H3K27ac peak.}
}
\value{
A data frame containing bulk H3K27ac peaks which are celltype-specific and
overlap > x\% of a celltype h3K27ac peak.
}
\description{
For a given celltype, this function takes as input the output of functions
BulkCelltypeOverlap() and CelltypeSpecificPeaks(), and will generate a data frame
containing bulk H3K27ac peaks which exclusively overlap the celltype investigated,
and each of these bulk H3K27ac peaks overlaps > x\% of a celltype
peak.
}
