% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neu_H3K27ac_hg38.R
\docType{data}
\name{neu_H3K27ac_hg38}
\alias{neu_H3K27ac_hg38}
\title{Neuronal H3K27ac peaks with hg38 coordinates.}
\format{
A data frame with 148055 rows and 4 variables:
\describe{
\item{CHR}{chromosome number}
\item{Start}{Start location of peak in BP}
\item{End}{End location of peak in BP}
\item{Peak}{Peak name}
...
}
}
\source{
\url{https://github.com/nottalexi/brain-cell-type-peak-files/tree/master/H3K27ac}
}
\usage{
neu_H3K27ac_hg38
}
\description{
A dataset containing neuronal H3K27ac peaks with hg38 coordinates in bed
file format.
}
\keyword{datasets}
