% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHAS.R
\name{plot_celltype_scores}
\alias{plot_celltype_scores}
\title{Plot cell type scores between two groups}
\usage{
plot_celltype_scores(celltypeScores, pheno)
}
\arguments{
\item{celltypeScores}{The output data frame from the function CelltypeScore()}

\item{pheno}{A data frame containing a column named Sample and a column named Group.}
}
\value{
A violin plot showing the cell type scores between two different groups.
}
\description{
This function takes as input the cell type scores generated using the function CelltypeScore()
and plots a violin plot of the cell type scores between two different groups.
}
