% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_correlation.R
\name{plot_correlation}
\alias{plot_correlation}
\title{Plot the correlation}
\usage{
plot_correlation(celltypeProportion, celltypeScores, group)
}
\arguments{
\item{celltypeProportion}{The output list from the function CelltypeProportion()}

\item{celltypeScores}{The output data frame from the function CelltypeScore()}

\item{group}{A data frame containing a column named "Sample" and a column named "Group".}
}
\value{
A plot showing the correlation for each cell type.
}
\description{
This function plots the correlation between matrix-factorisation-predicted cell type proportions and CHAS scores
}
