% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_celltype_props.R
\name{plot_MF_groups}
\alias{plot_MF_groups}
\title{Plot MF-predicted proportions between groups}
\usage{
plot_MF_groups(MFscores, group)
}
\arguments{
\item{MFscores}{The output data frame from the function CelltypeScore()}

\item{group}{A data frame containing a column named "Sample" and a column named "Group".}
}
\value{
A violin plot showing the cell type scores between two different groups.
}
\description{
This function creates a violin plot of the predicted cell type proportions,
split into cell types and groups of samples.
}
