% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_celltype_props_all.R
\name{plot_MF_scores}
\alias{plot_MF_scores}
\title{Plot matrix-factorisation-predicted cell type proportions in bulk samples}
\usage{
plot_MF_scores(MFscores)
}
\arguments{
\item{MFscores}{The output data frame from the function CelltypeProportion()}
}
\value{
A bar plot showing the cell type proportions for each sample.
}
\description{
This function creates a stacked bar plot containing the predicted cell-type proportions
for each bulk sample.
}
