% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReadCounting.R
\name{ReadCounting}
\alias{ReadCounting}
\title{Read Counting}
\usage{
ReadCounting(unionPeaksSaf, bulkDir, refDir, threads = 8)
}
\arguments{
\item{unionPeaksSaf}{Path to the SAF file containing the union peaks.}

\item{bulkDir}{Directory path where all the bulk BAM files are located.
The name of the input BAM files must only contain the sample ID, such as 'SRR5927824.bam' for
the bulk sample 'SRR5927824'}

\item{refDir}{Directory path where all the reference BAM files are located.
The name of the input BAM files must only contain the sample ID, such as 'astrocyte_1.bam' for
the reference sample 'astrocyte_1'.}

\item{threads}{The number of threads for running featureCounts. The default is 8.}
}
\value{
A list containing two dataframes
the first data frame contains the bulk counts
the second data frame contains the reference counts
}
\description{
This function counts the reads in bulk and reference bam files ar the union peaks.
}
\details{
This function takes three main inputs: the directory to where the bulk and reference bam files
are stored, and the directory to the .saf file.
}
