% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ref_H3K27ac_peaks_hg38.R
\docType{data}
\name{ref_H3K27ac_peaks_hg38}
\alias{ref_H3K27ac_peaks_hg38}
\title{Reference H3K27ac peaks with hg38 coordinates.}
\format{
A data frame with 162018 rows and 4 variables:
\describe{
\item{CHR}{chromosome number}
\item{Start}{Start location of peak in BP}
\item{End}{End location of peak in BP}
\item{Peak}{Peak name}
...
}
}
\source{
Generated using data from \url{https://github.com/nottalexi/brain-cell-type-peak-files/tree/master/H3K27ac}
}
\usage{
ref_H3K27ac_peaks_hg38
}
\description{
A dataset containing reference H3K27ac peaks from four basic brain cell types with hg38
coordinates in bed file format.
}
\keyword{datasets}
