% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_mf.R
\name{run_mf}
\alias{run_mf}
\title{Run Matrix Factorisation}
\usage{
run_mf(bulkCPM, ref_median, ref_var, signature)
}
\arguments{
\item{bulkCPM}{The bulk CPM data frame.}

\item{ref_median}{The reference median data frame.}

\item{ref_var}{The reference variability data frame.}

\item{signature}{The signature vector.}
}
\value{
The EPIC scores.
}
\description{
This function runs the matrix factorisation.
}
