% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/select_signature_peaks.R
\name{select_signature_peaks}
\alias{select_signature_peaks}
\title{Select Signature Peaks}
\usage{
select_signature_peaks(signature, median_df, num_samples)
}
\arguments{
\item{signature}{The signature vector.}

\item{median_df}{The median data frame.}

\item{num_samples}{The number of samples.}
}
\value{
A list containing the signature and count.
}
\description{
This function selects the signature peaks.
}
