% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_tpm.R
\name{calculate_cpm}
\alias{calculate_cpm}
\title{Calculate cpm}
\usage{
calculate_cpm(counts, peaks, signature)
}
\arguments{
\item{counts}{The counts data frame.}

\item{peaks}{The peaks data frame.}

\item{signature}{The signature vector.}
}
\value{
A data frame of cpm values.
}
\description{
This function calculates the Transcripts Per Million (cpm) for the given counts, peaks, and signature.
}
