% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHAS.R
\name{CelltypeScore}
\alias{CelltypeScore}
\title{Generate celltype-specific histone acetylation scores}
\usage{
CelltypeScore(cpm, celltypeSpecificPeaks)
}
\arguments{
\item{cpm}{A counts per million matrix with rows labeling peaks and columns labeling
samples.}

\item{celltypeSpecificPeaks}{The output list of dataframes from the function CelltypeSpecificPeaks() containing
the annotated bulk peaks and the cell type-specific bulk peaks.}
}
\value{
Celltype-specific histone acetylation scores for each sample.
}
\description{
This function takes as input a counts per million (cpm) matrix and the output of function
CelltypeSpecificPeaks(), and will calculate celltype-specific histone acetylation scores for
each sample in the cpm matrix.
}
