% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CelltypeScore.R
\name{CelltypeScore}
\alias{CelltypeScore}
\title{Generate cell-type-specific histone acetylation scores}
\usage{
CelltypeScore(counts, celltypeSpecificPeaks, method)
}
\arguments{
\item{counts}{A counts per million matrix with rows labeling peaks and columns labeling samples.}

\item{celltypeSpecificPeaks}{The output list of data frames from the function CelltypeSpecificPeaks() containing
the annotated bulk peaks and the cell type-specific bulk peaks.}

\item{method}{Whether mean or median should be used to calculate the score for each sample across cell type-specific peaks.}
}
\value{
Cell-type-specific histone acetylation scores for each sample.
}
\description{
This function takes as input a raw counts matrix and the output of function CelltypeSpecificPeaks(),
and will calculate cell-type-specific histone acetylation scores for each sample in the counts matrix.
}
