% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHAS.R
\name{CelltypeSpecificPeaks}
\alias{CelltypeSpecificPeaks}
\title{Identify celltype-specific H3K27ac peaks which overlap > x\% of the celltype}
\usage{
CelltypeSpecificPeaks(bulkPeaks, celltype, CelltypeOverlap, x)
}
\arguments{
\item{bulkPeaks}{A data frame of bulk H3K27ac peaks in BED format, with the first
column containing the chromosome number, the second column containing the start
position of the peak, the third column containing the end position of the peak, and
the fourth column containing the peak identifier.}

\item{celltype}{A character specifying the celltype of interest.}

\item{CelltypeOverlap}{output of the BulkCelltypeOverlap function for the celltype of interest.}

\item{x}{A number between 0 and 1 specifying the minimum overlap between the bulk peak and the celltype peak.}
}
\value{
A data frame containing bulk H3K27ac peaks which are celltype-specific.
}
\description{
This function identifies which bulk H3K27ac peaks are specific to a celltype peak,
and overlap >x\% of it.
}
