% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHAS.R
\name{CelltypeSpecificPeaks}
\alias{CelltypeSpecificPeaks}
\title{Identify cell type-specific peaks in bulk tissue H3K27ac profiles}
\usage{
CelltypeSpecificPeaks(bulkPeaks, celltypePeaks, p)
}
\arguments{
\item{bulkPeaks}{A data frame of bulk H3K27ac peaks in BED format, with the first
column containing the chromosome number, the second column containing the start
position of the peak, the third column containing the end position of the peak, and
the fourth column containing the peak identifier.}

\item{celltypePeaks}{A data frame of celltype H3K27ac peaks in BED format, with the
first column containing the chromosome number, the second column containing the start
position of the peak, the third column containing the end position of the peak, and
the fourth column containing the peak identifier.}

\item{p}{A number from 0-1 determining the percentage overlap required for a bulk peak
to be considered cell type specific.}
}
\value{
A data frame of annotated bulk peaks and a data frame containing only the cell type
specific bulk peaks.
}
\description{
This function takes as input bulk H3K27ac peaks and cell type sorted H3K27ac peaks,
and will generate a data frame with annotated bulk peaks and another data frame
made up of the cell type specific bulk peaks.
}
