% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ConsensusPeaks.R
\name{ConsensusPeaks}
\alias{ConsensusPeaks}
\title{Find Consensus Peaks}
\usage{
ConsensusPeaks(bulkPeaks, bulkCounts, refPeaks, refCounts, bedtools_path)
}
\arguments{
\item{bulkPeaks}{A data frame of bulk H3K27ac peaks in BED format
the first column contains the chromosome number following 'chr',
such as 'chr1'
the second column contains the start position of the peak
the third column contains the end position of the peak
the fourth column contains the peak identifier}

\item{bulkCounts}{A counts matrix for bulk data
rows represent peaks, using the same identifier as bulkPeaks
columns represent bulk samples}

\item{refPeaks}{A data frame of reference H3K27ac peaks in BED format
the first column contains the chromosome number, such as 'chr1'
the second column contains the start position of the peak
the third column contains the end position of the peak
the fourth column contains the peak identifier}

\item{refCounts}{A counts matrix for reference data
the rows represent peaks, using the same identifier as refPeaks
the columns represent cell-type samples, and can have one or multiple
samples for each cell type.}

\item{bedtools_path}{The path to where bedtools is installed
for example, in MacOS, this can be checked by running "\% which bedtools"
in Terminal.}
}
\value{
A list containing the following:
[1] data frame: bulk counts for consensus peaks
[2] data frame: reference counts for consensus peaks
}
\description{
This function identifies the consensus peaks and their counts
in each bulk and reference samples by switching IDs
}
\details{
This function takes four inputs:
bulk: peak location, raw counts
reference: peak location, raw counts

This function has three steps:
(1) find consensus peaks in bulk and reference
(2) assign new identifiers to consensus peaks
(3) filter the counts matrices to contain only the consensus peaks
}
