% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/CHAS.R, R/plot_celltype_props.R
\name{plot_celltype_props}
\alias{plot_celltype_props}
\title{Plot cell type proportions}
\usage{
plot_celltype_props(annotatedPeaks, exampleData = TRUE)

plot_celltype_props(annotatedPeaks, exampleData = TRUE)
}
\arguments{
\item{annotatedPeaks}{The output list of data frames from the function CelltypeSpecificPeaks()
containing the annotated bulk peaks and the cell type-specific bulk peaks.}
}
\value{
A stacked bar plot showing the proportion of each cell type in the bulk peak set.

A stacked bar plot showing the proportion of each cell type in the bulk peak set.
}
\description{
This function takes as input the output of the function CelltypeSpecificPeaks()
and uses the annotated bulk peaks to plot the proportion of each cell type.

This function takes as input the output of the function CelltypeSpecificPeaks()
and uses the annotated bulk peaks to plot the proportion of each cell type.
}
