% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/UnionPeaks.R
\name{UnionPeaks}
\alias{UnionPeaks}
\title{Find Union Peaks}
\usage{
UnionPeaks(bulkPeaks, refPeakList)
}
\arguments{
\item{bulkPeaks}{A data frame of bulk H3K27ac peaks in BED format
the first column contains the chromosome number, such as 'chr1'
the second column contains the start position of the peak
the third column contains the end position of the peak
the fourth column contains the peak identifier}

\item{refPeakList}{A list containing several data frames, each data frame contains the reference
H3K27ac peaks for one cell type in BED format
the first column contains the chromosome number, such as 'chr1'
the second column contains the start position of the peak
the third column contains the end position of the peak
the fourth column contains the peak identifier}
}
\value{
A list containing two data frames
the first data frame contains the consensus peaks
the second data frame contains the cell type-specific peaks
}
\description{
This function identifies the consensus peaks and assigns cell types to them.
}
